Split mobile recovery view state
