+++ Packages/iOS/CmuxMobileShellUI/Sources/CmuxMobileShellUI/SignInView.swift
@@ -10,10 +10,7 @@ import SwiftUI
 #elseif os(macOS)
 import AppKit
 #endif
-
 struct SignInView: View {
-    private static let authRestoreTimeout: Duration = .seconds(10)
-
     @Environment(AuthCoordinator.self) private var authManager
     @Environment(\.analytics) private var analytics
     @State private var email = ""
@@ -25,11 +22,8 @@ struct SignInView: View {
     @State private var shouldAutofocusCode = false
     @State private var shouldAutofocusEmail = false
     @State private var signInTask: Task<Void, Never>?
-    @State private var authRestoreTimedOut = false
-    @State private var authRestoreRetryGeneration = 0
     @FocusState private var isEmailFocused: Bool
     @FocusState private var isCodeFocused: Bool
-
     var body: some View {
         NavigationStack {
             ZStack {
@@ -42,17 +36,12 @@ struct SignInView: View {
                         UIApplication.shared.dismissMobileKeyboard()
                     }
                     .ignoresSafeArea()
-
                 VStack(spacing: 0) {
                     Spacer(minLength: 0)
-
                     signInEntrySwitcher
                 }
             }
             .mobileInlineNavigationTitle()
-            .task(id: authRestoreDeadlineTaskID) {
-                await updateAuthRestoreDeadline(isRestoring: authManager.isRestoringSession)
-            }
         }
     }
 
@@ -84,7 +73,7 @@ struct SignInView: View {
         authCard {
             VStack(spacing: 20) {
                 brandHeader
-                authRestoreStatus
+                SignInAuthRestoreStatusView()
 
                 Button {
                     signInTask = Task {
@@ -175,7 +164,7 @@ struct SignInView: View {
         authCard {
             VStack(spacing: 18) {
                 brandHeader
-                authRestoreStatus
+                SignInAuthRestoreStatusView()
 
                 VStack(spacing: 6) {
                     Text(L10n.string("mobile.signIn.checkEmail", defaultValue: "Check your email"))
@@ -250,84 +239,6 @@ struct SignInView: View {
                 .font(.subheadline)
                 .foregroundStyle(.secondary)
             }
-        }
-    }
-
-    @ViewBuilder
-    private var authRestoreStatus: some View {
-        if authManager.isRestoringSession {
-            VStack(alignment: .leading, spacing: 8) {
-                HStack(spacing: 8) {
-                    if authRestoreTimedOut {
-                        Image(systemName: "exclamationmark.triangle.fill")
-                            .foregroundStyle(.orange)
-                            .accessibilityHidden(true)
-                    } else {
-                        ProgressView()
-                            .controlSize(.small)
-                            .accessibilityHidden(true)
-                    }
-
-                    Text(authRestoreTimedOut
-                        ? L10n.string("mobile.signIn.restoreTimeoutTitle", defaultValue: "Still restoring session")
-                        : L10n.string("mobile.signIn.restoring", defaultValue: "Restoring session"))
-                        .font(.subheadline.weight(.semibold))
-                        .foregroundStyle(.primary)
-                }
-
-                Text(authRestoreTimedOut
-                    ? L10n.string(
-                        "mobile.signIn.restoreTimeoutMessage",
-                        defaultValue: "cmux could not finish checking your saved session. Check your connection, then retry."
-                    )
-                    : L10n.string(
-                        "mobile.signIn.restoringMessage",
-                        defaultValue: "Checking your saved session. Sign-in options are paused until this finishes."
-                    ))
-                    .font(.caption)
-                    .foregroundStyle(.secondary)
-                    .fixedSize(horizontal: false, vertical: true)
-
-                if authRestoreTimedOut {
-                    Button {
-                        retryAuthRestore()
-                    } label: {
-                        Text(L10n.string("mobile.common.retry", defaultValue: "Retry"))
-                    }
-                    .buttonStyle(.bordered)
-                    .controlSize(.small)
-                    .accessibilityIdentifier("signin.restoreRetry")
-                }
-            }
-            .frame(maxWidth: .infinity, alignment: .leading)
-            .accessibilityIdentifier("signin.restoreStatus")
-        }
-    }
-
-    private var authRestoreDeadlineTaskID: Int {
-        (authRestoreRetryGeneration &* 2) + (authManager.isRestoringSession ? 1 : 0)
-    }
-
-    private func updateAuthRestoreDeadline(isRestoring: Bool) async {
-        guard isRestoring else {
-            authRestoreTimedOut = false
-            return
-        }
-        authRestoreTimedOut = false
-        do {
-            try await ContinuousClock().sleep(for: Self.authRestoreTimeout)
-        } catch {
-            return
-        }
-        guard authManager.isRestoringSession else { return }
-        authRestoreTimedOut = true
-    }
-
-    private func retryAuthRestore() {
-        authRestoreTimedOut = false
-        authRestoreRetryGeneration &+= 1
-        Task {
-            await authManager.revalidateSession()
         }
     }
